+++ test/cloud/client.go
@@ -18,6 +18,7 @@ const defaultTestTimeout = 1 * time.Second
 // PutStack sets a new stack in the /v1/stacks/<org>/<stack id>.
 // Note: this is not a real endpoint.
 func PutStack(t *testing.T, addr string, orgUUID string, st cloud.StackResponse) {
+	t.Helper()
 	ctx, cancel := context.WithTimeout(context.Background(), defaultTestTimeout)
 	defer cancel()